split from one large file into individual components in separate files
